requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
attrs==25.2.0
beautifulsoup4==4.13.3
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.3.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fake-useragent==2.1.0
fastapi==0.115.11
h11
httplib2==0.22.0
httpx
idna==3.10
importlib_metadata==8.6.1
lxml==5.3.1
lxml_html_clean==0.4.1
numpy==2.2.3
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.3
parse==1.20.2
pillow
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2025.1
requests==2.32.3
requests-html==0.10.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.46.1
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
uvicorn
w3lib==2.3.1
websocket-client==1.8.0
zipp==3.21.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.1
pycountry==24.6.1
jellyfish
aiofiles
geopy
